chore: separated VC export, verification and presentation
